separated export modal and export notification and removed blocking export modal status view
